refactor(json-schema): extract type-specific helpers
