+++ lib/common.js
@@ -124,12 +124,6 @@ exports.isIsoDate = function (date) {
 };
 
 
-exports.isoDatePattern = function () {
-
-    return internals.isoDate.source.replace(/\\:/g, ':');
-};
-
-
 exports.isNumber = function (value) {
 
     return typeof value === 'number' && !isNaN(value);
