+++ lib/types/array.js
@@ -5,6 +5,7 @@ const { assert, deepEqual, reach } = require('@hapi/hoek');
 const Any = require('./any');
 const Common = require('../common');
 const Compile = require('../compile');
+const JsonArray = require('../json-schema/array');
 
 
 const internals = {};
@@ -70,85 +71,7 @@ module.exports = Any.extend({
 
     jsonSchema(schema, res, mode, options) {
 
-        const ordered = schema.$_terms.ordered;
-
-        // Handle ordered items (tuple-like) using 'prefixItems'
-
-        if (ordered.length) {
-            res.prefixItems = ordered.map((item) => item.$_jsonSchema(mode, options));
-        }
-
-        if (schema.$_terms.items.length) {
-            let items;
-            if (schema.$_terms.items.length === 1) {
-                items = schema.$_terms.items[0].$_jsonSchema(mode, options);
-            }
-            else {
-                items = {
-                    anyOf: schema.$_terms.items.map((item) => item.$_jsonSchema(mode, options))
-                };
-            }
-
-            // If there are ordered items, remaining items are 'unevaluatedItems'
-
-            if (ordered.length) {
-                res.unevaluatedItems = items;
-                internals.setOrderedMinItems(res, ordered);
-            }
-            else {
-                res.items = items;
-            }
-        }
-        else if (ordered.length) {
-            // No additional items allowed beyond the ordered ones
-
-            res.unevaluatedItems = false;
-            internals.setOrderedMinItems(res, ordered);
-            res.maxItems = ordered.length;
-        }
-
-        // Map 'has' rules to 'contains' in JSON Schema
-
-        const contains = [];
-        for (const rule of schema._rules) {
-            if (rule.name === 'has' &&
-                !rule.args.schema._refs.refs.length) {
-
-                contains.push(rule.args.schema.$_jsonSchema(mode, options));
-            }
-        }
-
-        if (contains.length) {
-            if (contains.length === 1) {
-                res.contains = contains[0];
-            }
-            else {
-                res.allOf = contains.map((item) => ({ contains: item }));
-            }
-        }
-
-        if (schema._flags.single &&
-            schema.$_terms.items.length) {
-
-            let items;
-            if (schema.$_terms.items.length === 1) {
-                items = schema.$_terms.items[0].$_jsonSchema(mode, options);
-            }
-            else {
-                items = {
-                    anyOf: schema.$_terms.items.map((item) => item.$_jsonSchema(mode, options))
-                };
-            }
-
-            res = {
-                anyOf: [
-                    res,
-                    items
-                ]
-            };
-        }
-
-        return res;
+        return JsonArray.emit(schema, res, mode, options);
     },
 
     rules: {
@@ -768,31 +691,6 @@ module.exports = Any.extend({
         'array.unique': '{{#label}} contains a duplicate value'
     }
 });
-
-
-// Helpers
-
-internals.setOrderedMinItems = function (res, ordered) {
-
-    // Ordered items are optional by default; the array only needs to reach the
-    // last explicitly required position.
-    const minItems = internals.orderedMinItems(ordered);
-    if (minItems) {
-        res.minItems = minItems;
-    }
-};
-
-
-internals.orderedMinItems = function (ordered) {
-
-    for (let i = ordered.length - 1; i >= 0; --i) {
-        if (ordered[i]._flags.presence === 'required') {
-            return i + 1;
-        }
-    }
-
-    return 0;
-};
 
 
 internals.fillMissedErrors = function (schema, errors, requireds, value, state, prefs) {
