+++ lib/types/date.js
@@ -4,13 +4,12 @@ const { assert } = require('@hapi/hoek');
 
 const Any = require('./any');
 const Common = require('../common');
+const JsonDate = require('../json-schema/date');
 const Template = require('../template');
 
 
 const internals = {
-    formats: ['iso', 'javascript', 'unix'],
-    maxJs: 100e6 * 24 * 60 * 60 * 1000,        // 100 million days in ms (ECMA-262 §21.4.1.1)
-    maxUnix: 100e6 * 24 * 60 * 60              // 100 million days in seconds
+    formats: ['iso', 'javascript', 'unix']
 };
 
 
@@ -54,34 +53,7 @@ module.exports = Any.extend({
 
     jsonSchema(schema, res, mode, options) {
 
-        const format = schema._flags.format;
-
-        if (format === 'javascript') {
-            res.type = 'number';
-            res.minimum = -internals.maxJs;
-            res.maximum = internals.maxJs;
-            return res;
-        }
-
-        if (format === 'unix') {
-            res.type = 'number';
-            res.minimum = -internals.maxUnix;
-            res.maximum = internals.maxUnix;
-            return res;
-        }
-
-        if (format === 'iso') {
-            res.type = 'string';
-            res.pattern = Common.isoDatePattern();
-            return res;
-        }
-
-        res.type = ['string', 'number'];
-        res.format = 'date-time';
-        res.minimum = -internals.maxJs;
-        res.maximum = internals.maxJs;
-
-        return res;
+        return JsonDate.emit(schema, res);
     },
 
     rules: {
@@ -127,12 +99,7 @@ module.exports = Any.extend({
             },
             jsonSchema(rule, res) {
 
-                const date = rule.args.date;
-                if (date instanceof Date) {
-                    res['x-constraint'] = { ...res['x-constraint'], greater: date.toISOString() };
-                }
-
-                return res;
+                return JsonDate.appendConstraint(res, 'greater', rule.args.date);
             }
         },
 
@@ -150,12 +117,7 @@ module.exports = Any.extend({
             },
             jsonSchema(rule, res) {
 
-                const date = rule.args.date;
-                if (date instanceof Date) {
-                    res['x-constraint'] = { ...res['x-constraint'], less: date.toISOString() };
-                }
-
-                return res;
+                return JsonDate.appendConstraint(res, 'less', rule.args.date);
             }
         },
 
@@ -166,12 +128,7 @@ module.exports = Any.extend({
             },
             jsonSchema(rule, res) {
 
-                const date = rule.args.date;
-                if (date instanceof Date) {
-                    res['x-constraint'] = { ...res['x-constraint'], max: date.toISOString() };
-                }
-
-                return res;
+                return JsonDate.appendConstraint(res, 'max', rule.args.date);
             }
         },
 
@@ -182,12 +139,7 @@ module.exports = Any.extend({
             },
             jsonSchema(rule, res) {
 
-                const date = rule.args.date;
-                if (date instanceof Date) {
-                    res['x-constraint'] = { ...res['x-constraint'], min: date.toISOString() };
-                }
-
-                return res;
+                return JsonDate.appendConstraint(res, 'min', rule.args.date);
             }
         },
