+++ lib/types/keys.js
@@ -698,6 +698,9 @@ module.exports = Any.extend({
     }
 });
 
+
+// Helpers
+
 internals.clone = function (value, prefs) {
 
     // Object
