+++ lib/types/string.js
@@ -1,12 +1,12 @@
 'use strict';
 
-const Url = require('url');
 const { assert, escapeRegex } = require('@hapi/hoek');
 const { isDomainValid, isEmailValid, ipRegex, uriRegex } = require('@hapi/address');
 const Tlds = require('@hapi/tlds');
 
 const Any = require('./any');
 const Common = require('../common');
+const JsonString = require('../json-schema/string');
 
 
 const internals = {
@@ -148,19 +148,7 @@ module.exports = Any.extend({
 
     jsonSchema(schema, res, mode, options) {
 
-        const noEmpty = !schema._valids?.has('') && !schema._flags.only;
-        if (noEmpty) {
-            const min = schema.$_getRule('min');
-            const length = schema.$_getRule('length');
-
-            if ((!min || min._resolve.length || min.args.limit > 0) &&
-                (!length || length._resolve.length || length.args.limit > 0)) {
-
-                res.minLength = 1;
-            }
-        }
-
-        return res;
+        return JsonString.applyBase(schema, res);
     },
 
     rules: {
@@ -180,7 +168,7 @@ module.exports = Any.extend({
             },
             jsonSchema(rule, res) {
 
-                return internals.appendPattern(res, '^[a-zA-Z0-9]+$');
+                return JsonString.appendPattern(res, '^[a-zA-Z0-9]+$');
             }
         },
 
@@ -206,7 +194,7 @@ module.exports = Any.extend({
             },
             jsonSchema(rule, res) {
 
-                return internals.appendPattern(res, internals.base64JsonSchemaPattern(rule.args.options));
+                return JsonString.appendPattern(res, JsonString.base64Pattern(rule.args.options, internals.base64Regex));
             }
         },
 
@@ -290,7 +278,7 @@ module.exports = Any.extend({
             },
             jsonSchema(rule, res) {
 
-                return internals.appendPattern(res, internals.dataUriJsonSchemaPattern(rule.args.options));
+                return JsonString.appendPattern(res, JsonString.dataUriPattern(rule.args.options, internals.base64Regex));
             }
         },
 
@@ -317,7 +305,7 @@ module.exports = Any.extend({
                 // Not using format: 'hostname' because it accepts single-label names (e.g. 'localhost')
                 // while Joi's domain() requires at least 2 segments by default
 
-                return internals.appendPattern(res, internals.domainJsonSchemaPattern(rule.address));
+                return JsonString.appendPattern(res, JsonString.domainPattern(rule.address, internals.minDomainSegments));
             }
         },
 
@@ -481,7 +469,7 @@ module.exports = Any.extend({
             },
             jsonSchema(rule, res, isOnly, mode) {
 
-                return internals.appendPattern(res, internals.hexJsonSchemaPattern(rule.args.options, mode));
+                return JsonString.appendPattern(res, JsonString.hexPattern(rule.args.options, mode));
             }
         },
 
@@ -502,7 +490,7 @@ module.exports = Any.extend({
             },
             jsonSchema(rule, res) {
 
-                return internals.appendPattern(res, internals.hostnameJsonSchemaPattern());
+                return JsonString.appendPattern(res, JsonString.hostnamePattern(internals.minDomainSegments));
             }
         },
 
@@ -536,7 +524,7 @@ module.exports = Any.extend({
             },
             jsonSchema(rule, res) {
 
-                return internals.appendPattern(res, internals.ipJsonSchemaPattern(rule.args.options));
+                return JsonString.appendPattern(res, JsonString.ipPattern(rule.args.options));
             }
         },
 
@@ -694,7 +682,7 @@ module.exports = Any.extend({
             },
             jsonSchema(rule, res) {
 
-                return internals.appendPattern(res, rule.args.regex.source);
+                return JsonString.appendPattern(res, rule.args.regex.source);
             },
             args: ['regex', 'options'],
             multi: true
@@ -736,7 +724,7 @@ module.exports = Any.extend({
             },
             jsonSchema(rule, res) {
 
-                return internals.appendPattern(res, '^[A-Za-z0-9_]+$');
+                return JsonString.appendPattern(res, '^[A-Za-z0-9_]+$');
             }
         },
 
@@ -955,287 +943,6 @@ internals.validateTlds = function (set, source) {
     for (const tld of set) {
         assert(isDomainValid(tld, { minDomainSegments: 1, maxDomainSegments: 1 }), `${source} must contain valid top level domain names`);
     }
-};
-
-
-internals.base64JsonSchemaPattern = function (options) {
-
-    return internals.base64Regex[options.paddingRequired][options.urlSafe].source;
-};
-
-
-internals.appendPattern = function (res, pattern) {
-
-    if (res.allOf) {
-        const existingPattern = res.pattern;
-        delete res.pattern;
-        res.allOf.push(...[existingPattern].filter((value) => value !== undefined).map((value) => ({ pattern: value })), { pattern });
-        return res;
-    }
-
-    if (res.pattern === undefined) {
-        res.pattern = pattern;
-        return res;
-    }
-
-    res.allOf = [
-        { pattern: res.pattern },
-        { pattern }
-    ];
-
-    delete res.pattern;
-    return res;
-};
-
-
-internals.hostnameJsonSchemaPattern = function () {
-
-    const hostname = internals.unanchoredPattern(internals.domainJsonSchemaPattern({ minDomainSegments: 1, tlds: false }));
-    const ip = internals.unanchoredPattern(internals.ipJsonSchemaPattern({ cidr: 'forbidden' }));
-
-    return `^(?:${hostname}|${ip})$`;
-};
-
-
-internals.ipJsonSchemaPattern = function (options) {
-
-    return ipRegex(options).regex.source.replace(/\[\\w-\\\./g, '[\\w.\\-');
-};
-
-
-internals.dataUriJsonSchemaPattern = function (options) {
-
-    const mediaType = 'data:[\\w+.-]+\\/[\\w+.-]+;';
-    const base64 = internals.unanchoredPattern(internals.base64Regex[options.paddingRequired].false);
-
-    return `^${mediaType}(?:base64,${base64}|(?!base64,).*)$`;
-};
-
-
-internals.hexJsonSchemaPattern = function (options, mode) {
-
-    const digits = '[0-9A-Fa-f]+';
-    const bytes = '(?:[0-9A-Fa-f]{2})+';
-
-    if (!options.byteAligned) {
-        if (options.prefix === true) {
-            return '^0[xX][0-9A-Fa-f]+$';
-        }
-
-        if (options.prefix === 'optional') {
-            return '^(?:0[xX])?[0-9A-Fa-f]+$';
-        }
-
-        return '^[0-9A-Fa-f]+$';
-    }
-
-    if (options.prefix === true) {
-        return `^0[xX]${bytes}$`;
-    }
-
-    if (options.prefix === 'optional') {
-        if (mode === 'output') {
-            return `^(?:${bytes}|0[xX]${bytes})$`;
-        }
-
-        return `^(?:${digits}|0[xX]${bytes})$`;
-    }
-
-    if (mode === 'output') {
-        return `^${bytes}$`;
-    }
-
-    return '^[0-9A-Fa-f]+$';
-};
-
-
-internals.unanchoredPattern = function (regex) {
-
-    const pattern = typeof regex === 'string' ? regex : regex.source;
-    return pattern.replace(/^\^/, '').replace(/\$$/, '');
-};
-
-
-internals.domainJsonSchemaPattern = function (options) {
-
-    const settings = internals.addressOptions(options);
-
-    const min = (settings.minDomainSegments || internals.minDomainSegments) - 1;
-    const max = settings.maxDomainSegments !== undefined ? settings.maxDomainSegments - 1 : '';
-    const fqdn = settings.allowFullyQualified ? '\\.?' : '';
-    const totalLength = '(?=.{1,256}$)';
-    const labelLength = '(?=[^.]{1,63}\\.)';
-    const tldLength = `(?=[^.]{1,63}${settings.allowFullyQualified ? '(?:\\.?$)' : '$'})`;
-    const label = internals.domainSegmentPattern({
-        allowUnicode: settings.allowUnicode !== false,
-        allowUnderscore: settings.allowUnderscore
-    });
-    const tld = internals.domainTldPattern(settings);
-    const denied = internals.domainDeniedTldLookahead(settings);
-
-    return `^${totalLength}(?:${labelLength}${label}\\.){${min},${max}}${denied}${tldLength}${tld}${fqdn}$`;
-};
-
-
-internals.domainSegmentPattern = function ({ allowUnicode, allowUnderscore, tld = false }) {
-
-    const nonAscii = allowUnicode ? '\\u0080-\\u{10FFFF}' : '';
-    const start = `[${tld ? 'A-Za-z' : `A-Za-z0-9${allowUnderscore ? '_' : ''}`}${nonAscii}]`;
-    const body = `[A-Za-z0-9${nonAscii}-]`;
-    const end = `[A-Za-z0-9${nonAscii}]`;
-
-    return `${start}(?:${body}*${end})?`;
-};
-
-
-internals.domainTldPattern = function (options = {}) {
-
-    const allow = internals.tldPatternValues(options.tlds && options.tlds.allow, options.allowUnicode !== false);
-    if (allow !== null) {
-        if (!allow.length) {
-            return '(?!)';
-        }
-
-        return `(?:${allow.join('|')})`;
-    }
-
-    return internals.domainSegmentPattern({
-        allowUnicode: options.allowUnicode !== false,
-        tld: true
-    });
-};
-
-
-internals.domainDeniedTldLookahead = function (options = {}) {
-
-    const deny = internals.tldPatternValues(options.tlds && options.tlds.deny, options.allowUnicode !== false);
-    if (!deny ||
-        !deny.length) {
-
-        return '';
-    }
-
-    const denied = deny.join('|');
-    const suffix = options.allowFullyQualified ? '\\.?$' : '$';
-    return `(?!(?:${denied})${suffix})`;
-};
-
-internals.tldValues = function (values) {
-
-    if (values instanceof Set) {
-        return [...values];
-    }
-
-    return null;
-};
-
-
-internals.tldPatternValues = function (values, allowUnicode) {
-
-    const tlds = internals.tldValues(values);
-    if (!tlds) {
-        return null;
-    }
-
-    const patterns = new Set();
-    for (const tld of tlds) {
-        const canonical = internals.domainTldValue(tld);
-        if (!canonical) {
-            continue;
-        }
-
-        patterns.add(internals.caseInsensitiveLiteral(canonical));
-
-        if (!allowUnicode) {
-            continue;
-        }
-
-        const unicode = Url.domainToUnicode(canonical).normalize('NFC');
-        if (unicode &&
-            unicode !== canonical) {
-
-            for (const variant of internals.unicodeTldVariants(canonical, unicode)) {
-                patterns.add(internals.caseInsensitiveLiteral(variant));
-            }
-        }
-    }
-
-    return [...patterns].sort();
-};
-
-
-internals.unicodeTldVariants = function (canonical, unicode) {
-
-    const variants = new Set([
-        unicode,
-        unicode.normalize('NFD'),
-        unicode.toLowerCase().normalize('NFC'),
-        unicode.toUpperCase().normalize('NFC')
-    ]);
-
-    const valid = [];
-    for (const variant of variants) {
-        if (Url.domainToASCII(variant).toLowerCase() === canonical) {
-            valid.push(variant);
-        }
-    }
-
-    return valid;
-};
-
-
-internals.domainTldValue = function (value) {
-
-    if (/[^\x00-\x7f]/.test(value)) {
-
-        return null;
-    }
-
-    const lower = value.toLowerCase();
-    if (value !== lower) {
-        return null;
-    }
-
-    return lower;
-};
-
-
-internals.caseInsensitiveLiteral = function (value) {
-
-    const pattern = [];
-    for (const char of value) {
-        const lower = char.toLowerCase();
-        const upper = char.toUpperCase();
-
-        if (lower.length !== 1 ||
-            upper.length !== 1) {
-
-            pattern.push(internals.regexLiteral(char));
-            continue;
-        }
-
-        if (lower === upper) {
-
-            pattern.push(internals.regexLiteral(char));
-            continue;
-        }
-
-        pattern.push(`[${internals.regexClassLiteral(lower)}${internals.regexClassLiteral(upper)}]`);
-    }
-
-    return pattern.join('');
-};
-
-
-internals.regexClassLiteral = function (value) {
-
-    return value.replace(/[\\\]\[\^-]/g, '\\$&');
-};
-
-
-internals.regexLiteral = function (value) {
-
-    return value.replace(/[|\\{}()[\]^$+*?.]/g, '\\$&');
 };
